+++ ft_strlcpy.c
@@ -6,16 +6,16 @@
 /*   By: skhaliff <skhaliff@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/07 15:03:20 by skhaliff          #+#    #+#             */
-/*   Updated: 2022/07/23 11:35:32 by skhaliff         ###   ########.fr       */
+/*   Updated: 2022/08/03 11:21:49 by skhaliff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"so_long.h"
 
-size_t	ft_strlcpy(char *dst, char *src, size_t dstsize)
+int	ft_strlcpy(char *dst, char *src, int dstsize)
 {
-	size_t	res;
-	size_t	i;
+	int	res;
+	int	i;
 
 	res = ft_strlen(src);
 	i = 0;
